bot/requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
arrow==1.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
blessed==1.20.0
certifi==2024.2.2
charset-normalizer==3.3.2
Deprecated==1.2.14
distro==1.9.0
Django==5.0.2
django-picklefield==3.1
django-telegrambot==1.0.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
netaddr==0.10.1
numpy==1.26.4
openpyxl==3.1.2
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.1
python-telegram-bot==20.8
pytz==2024.1
redis==3.5.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
yookassa==3.0.1